delelte service, serviceType, update serviceType, xem service vs article

+++ Dashboard/dental-dashboard/src/pages/ManageService.jsx
@@ -19,16 +19,19 @@ import {
   Typography,
   TablePagination,
 } from "@mui/material";
-import axios from "../config/axiosConfig";
 import { Visibility, Delete, Edit, Add, EditNote } from "@mui/icons-material";
+import { useNavigate } from 'react-router-dom';
+
 import Cookies from "js-cookie";
 import useGetAllService from "../hooks/service/useGetAllServiceType";
 import { jwtDecode } from "jwt-decode";
 import useUserStore from "../hooks/auth/useUserStore";
 import CreateServiceType from "../components/ManageService/createServiceType";
 import CreateService from "../components/ManageService/createService";
-import UpdateService from "../components/ManageService/updateService";
-
+import ServiceDetailModal from "../components/ManageService/ServiceDetail";
+import axios from "../config/axiosConfig";
+import EditService from "../components/ManageService/editService";
+import EditArticleService from "../components/ManageService/editArticleService";
 const units = {
   tooth: "Răng",
   jaw: "Hàm",
@@ -38,7 +41,11 @@ const units = {
 };
 
 const ServiceManagement = () => {
-
+  const navigate = useNavigate();
+  const [openModal, setOpenModal] = useState(false);
+  const [openEditService, setOpenEditService] = useState(false); 
+  const [openEditArticle, setOpenEditArticle] = useState(false);
+  const [serviceId, setServiceId] = useState(null);
   const [searchOption, setSearchOption] = useState("category");
   const [searchTerm, setSearchTerm] = useState("");
   const [page, setPage] = useState(0);
@@ -47,8 +54,6 @@ const ServiceManagement = () => {
   const { userLoggedIn, setUserLoggedIn, token } = useUserStore();
   const [openCreateServiceType, setOpenCreateServiceType] = useState(false);
   const [openCreateService, setOpenCreateService] = useState(false);
-const [openUpdateService, setOpenUpdateService]= useState(false);
-const [serviceToEdit, setServiceToEdit] = useState(null);
 
   const handleSearchChange = (event) => setSearchTerm(event.target.value);
   const handleChangePage = (event, newPage) => setPage(newPage);
@@ -101,8 +106,12 @@ const [serviceToEdit, setServiceToEdit] = useState(null);
       getAllService();
     }
   };
-  const handleDeleteService = async (serviceId) => {
+  const handleDeleteService = async (serviceId, typeId) => {
+    if (!serviceId || !typeId) {
+      console.error("Thiếu Id dịch vụ hoặc Id loại dịch vụ");
+      return;
 
+    }
     try {
       // Gọi API xóa loại dịch vụ, ví dụ như:
       const response = await axios.delete(`/service/delete/${serviceId}`,
@@ -112,25 +121,44 @@ const [serviceToEdit, setServiceToEdit] = useState(null);
           }
         }
       );
-
-      console.log("Xóa loại dịch vụ thành công:", response.data);
-
-
+      console.log("Xóa loại dịch vụ thành công:", response.data.message);
+      // Thực hiện cập nhật lại UI hoặc dữ liệu sau khi xóa thành công
+      handleRefreshServices();
     } catch (error) {
-      console.error("Lỗi khi xóa loại dịch vụ:", error);
+      console.error("Lỗi khi xóa loại dịch vụ:", error.response?.data?.message || error.message);
     }
 
   };
-// cập nhật dữ liệu cho service 
-const handleOpenUpdateService = (service) => {
-  setServiceToEdit(service);
-  setOpenUpdateService(true);
-};
 
-const handleCloseUpdateService = () => {
-  setOpenUpdateService(false);
-  setServiceToEdit(null);
-};
+  // Mở modal khi bấm vào icon "Xem chi tiết"
+  const handleOpenModal = (id) => {
+    console.log("id hiện tại là:", id);
+
+    setServiceId(id);  // Lưu ID của dịch vụ
+    setOpenModal(true);  // Mở modal
+  };
+  
+  // Đóng modal
+  const handleCloseModal = () => {
+    setOpenModal(false);
+  };
+
+  const handleOpenEditService = (id) => {
+    setServiceId(id);
+    setOpenEditService(true);  // Mở modal chỉnh sửa dịch vụ
+  };
+
+  const handleOpenEditArticle = (id) => {
+    setServiceId(id);
+    setOpenEditArticle(true);  // Mở modal chỉnh sửa bài viết dịch vụ
+  };
+  const handleCloseEditService = () => {
+    setOpenEditService(false);  // Đóng modal chỉnh sửa dịch vụ
+  };
+
+  const handleCloseEditArticle = () => {
+    setOpenEditArticle(false);  // Đóng modal chỉnh sửa bài viết
+  };
   return (
     <Box sx={{ paddingY: 6, paddingX: 0.5 }}>
       <Typography variant="h6" sx={{ fontWeight: "bold", mb: 2 }}>
@@ -204,12 +232,18 @@ const handleCloseUpdateService = () => {
             >
               Thêm mới loại dịch vụ
             </Button>
-            <Button variant="contained" color="primary" startIcon={<Edit />}>
+
+            <Button variant="contained" color="primary" startIcon={<Edit />}
+              onClick={() => navigate("/dashboard/quan-ly-loai-dich-vu")}>
               Chỉnh sửa loại dịch vụ
             </Button>
-            <Button variant="contained" color="error" startIcon={<Delete />}>
+
+            {/* <Button variant="contained" color="error" startIcon={<Delete />}
+              onClick={() => handleDeleteServiceType(category._id)}
+            >
+
               Xóa loại dịch vụ
-            </Button>
+            </Button> */}
           </Box>
           <Button
             variant="contained"
@@ -303,7 +337,9 @@ const handleCloseUpdateService = () => {
                       </TableCell>
                       <TableCell>
                         <Tooltip title="Xem chi tiết">
-                          <IconButton sx={{ color: "#1976d2" }}>
+                          <IconButton sx={{ color: "#1976d2" }}
+                            onClick={() => handleOpenModal(service._id)}
+                          >
                             <Visibility />
                           </IconButton>
                         </Tooltip>
@@ -312,23 +348,21 @@ const handleCloseUpdateService = () => {
                           <>
                             <Tooltip title="Chỉnh sửa dịch vụ">
                               <IconButton sx={{ color: "#4caf50" }}
-                              onClick={() => handleOpenUpdateService(service)}
-                              >
+                              onClick={() => handleOpenEditService(service._id)}>
                                 <Edit />
                               </IconButton>
                             </Tooltip>
-
                             <Tooltip title="Xóa dịch vụ">
                               <IconButton
                                 sx={{ color: "#f44336" }}
-                                onClick={() => handleDeleteService(service._id)}
+                                onClick={() => handleDeleteService(service._id, category._id)}
                               >
                                 <Delete />
                               </IconButton>
-
                             </Tooltip>
                             <Tooltip title="Chỉnh sửa bài viết dịch vụ">
-                              <IconButton sx={{ color: "#ff9800" }}>
+                              <IconButton sx={{ color: "#ff9800" }} 
+                              onClick={() => handleOpenEditArticle(service._id)}>
                                 <EditNote />
                               </IconButton>
                             </Tooltip>
@@ -363,14 +397,27 @@ const handleCloseUpdateService = () => {
         onClose={handleCloseCreateService}
         onRefresh={handleRefreshServices}
       />
-     {openUpdateService && serviceToEdit && (
-      <UpdateService
-        open={openUpdateService}
-        onClose={handleCloseUpdateService}
-        service={serviceToEdit} // Truyền dữ liệu dịch vụ vào form
-        onRefresh={handleRefreshServices}
+      <ServiceDetailModal
+        serviceId={serviceId}  // Truyền ID dịch vụ vào modal
+        open={openModal}  // Điều khiển trạng thái mở của modal
+        closeModal={handleCloseModal}  // Hàm đóng modal
+
+
       />
-    )}
+       {/* Modal chỉnh sửa dịch vụ */}
+       <EditService
+       serviceId={serviceId}
+        open={openEditService}
+        onClose={handleCloseEditService}
+      />
+
+      {/* Modal chỉnh sửa bài viết dịch vụ */}
+      <EditArticleService
+      serviceId={serviceId}
+        open={openEditArticle}
+        onClose={handleCloseEditArticle}
+      />
+      
     </Box>
   );
 };
